dev/service-plus-server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
ariadne
psycopg[binary]
websockets
PyJWT
bcrypt
